refactor iNMF
